+++ pennylane/ops/qubit/parametric_ops_multi_qubit.py
@@ -26,8 +26,13 @@ import numpy as np
 import pennylane as qml
 from pennylane import math, queuing
 from pennylane.capture.autograph import disable_autograph
-from pennylane.decomposition import add_decomps, controlled_resource_rep, register_resources
-from pennylane.decomposition.resources import register_reconstructor, resource_rep
+from pennylane.decomposition import (
+    add_decomps,
+    controlled_resource_rep,
+    register_resources,
+    resource_rep,
+)
+from pennylane.decomposition.reconstruct import register_reconstructor
 from pennylane.decomposition.symbolic_decomposition import adjoint_rotation, pow_rotation
 from pennylane.math.decomposition import decomp_int_to_powers_of_two
 from pennylane.operation import FlatPytree, Operation, Operator
